rename from fix_robustness to __init__.py
